BUG: Occasionally ImageMath Test 9 fails - morphological erode
Results are not as expected occasionally on select platforms.  Results
are inconsistent (sometimes perfect, other times with small deviations).

Using radius as int instead of float, to see if rounding/float errors
are to blame.

+++ TubeTKLib/Filtering/tubeImageMathFilters.h
@@ -103,7 +103,7 @@ public:
 
   /** Mathematical morphology using a sphere. Mode: 0=erode, 1=dilate. */
   static void MorphImage( typename ImageType::Pointer & imIn, int mode,
-    float radius, float foregroundValue, float backgroundValue );
+    int radius, float foregroundValue, float backgroundValue );
 
   /** Replace values within the image, with a mask. */
   static void OverwriteImage( typename ImageType::Pointer imIn,
